Refactor data view preference handling and update related URLs
