+++ bloomerp/django_bloomerp/bloomerp/components/objects/dataviews/dataview.py
@@ -3,9 +3,8 @@ import json
 
 from django.forms import modelform_factory
 from django import forms
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import render
 from django.urls import reverse
-from django.middleware.csrf import get_token
 from django.core.exceptions import FieldDoesNotExist
 from bloomerp.components.application_fields.filters import filters_init
 from bloomerp.forms.auth import User
@@ -20,7 +19,6 @@ from django.contrib.contenttypes.models import ContentType
 from bloomerp.services.permission_services import UserPermissionManager
 from bloomerp.services.permission_services import create_permission_str
 from bloomerp.services.user_services import get_data_view_fields
-from bloomerp.services.user_services import toggle_field_visibility
 from bloomerp.services.object_services import string_search_on_queryset
 from bloomerp.utils.filters import filter_model
 from bloomerp.models.users.user_list_view_preference import UserListViewPreference, ViewTypeEnum
@@ -297,130 +295,6 @@ def _get_data_view_options_form(
     return form_cls(initial=_get_data_view_options_initial(preference, definition.key))
 
 
-def _render_display_options(
-    request: HttpRequest,
-    content_type_id: int,
-    preference: UserListViewPreference,
-) -> HttpResponse:
-    data_view_fields = get_data_view_fields(preference)
-    return render(
-        request,
-        "cotton/features/dataviews/display_options.html",
-        {
-            "content_type_id": content_type_id,
-            "view_types": [vt.value for vt in ViewTypeEnum],
-            "preference": preference,
-            "fields": data_view_fields,
-            "accessible_fields": _get_accessible_application_fields(data_view_fields),
-            "dataview_options_form": _get_data_view_options_form(
-                preference,
-                _get_accessible_application_fields(data_view_fields),
-                request,
-            ),
-            "csrf_token": get_token(request),
-        },
-    )
-
-
-def _get_preference_operation(post_data) -> str | None:
-    if "view_type" in post_data:
-        return "change_type"
-    if "split_view_enabled" in post_data:
-        return "split_view"
-    if "dataview_options_view_type" in post_data:
-        return "opt"
-    if "toggle_field_id" in post_data:
-        return "field"
-    if "default_filters" in post_data:
-        return "default_filters"
-    return None
-
-
-def _change_default_filters(preference: UserListViewPreference, post_data) -> HttpResponse | None:
-    try:
-        payload = json.loads(post_data.get("default_filters") or "{}")
-    except json.JSONDecodeError:
-        return HttpResponse("Invalid default filters", status=400)
-
-    preference.default_filters = _normalize_default_filters(payload)
-    preference.save(update_fields=["default_filters"])
-    return None
-
-
-def _change_data_view_type(preference: UserListViewPreference, post_data) -> HttpResponse | None:
-    view_type = post_data["view_type"]
-    if _get_data_view_type_definition(view_type) is None:
-        return HttpResponse("Invalid view type", status=400)
-
-    preference.view_type = view_type
-    preference.save(update_fields=["view_type"])
-    return None
-
-
-def _change_split_view(preference: UserListViewPreference, post_data) -> HttpResponse | None:
-    preference.split_view_enabled = str(post_data["split_view_enabled"]).lower() == "true"
-    preference.save(update_fields=["split_view_enabled"])
-    return None
-
-
-def _change_data_view_options(
-    preference: UserListViewPreference,
-    data_view_fields,
-    post_data,
-) -> HttpResponse | None:
-    view_type = post_data["dataview_options_view_type"]
-    if view_type != preference.view_type:
-        return HttpResponse("Invalid options view type", status=400)
-
-    definition = _get_data_view_type_definition(view_type)
-    if definition is None:
-        return HttpResponse("Invalid view type", status=400)
-
-    form_cls = definition.create_opts_form(data_view_fields)
-    form = form_cls(post_data)
-    if not form.is_valid():
-        return HttpResponse("Invalid options", status=400)
-
-    options = dict(preference.options or {})
-    option_model = definition.get_options_model()
-    try:
-        options[view_type] = option_model.model_validate(form.cleaned_data).model_dump()
-    except PydanticValidationError as error:
-        return HttpResponse(f"Invalid options: {error}", status=400)
-
-    preference.options = options
-    preference.save(update_fields=["options"])
-    return None
-
-
-def _change_data_view_field_visibility(
-    request: HttpRequest,
-    content_type: ContentType,
-    preference: UserListViewPreference,
-    post_data,
-) -> HttpResponse | None:
-    try:
-        field_id = int(post_data["toggle_field_id"])
-        view_type = post_data.get("toggle_view_type", preference.view_type)
-        if _get_data_view_type_definition(view_type) is None:
-            return HttpResponse("Invalid view type", status=400)
-
-        permission_manager = UserPermissionManager(request.user)
-        application_field = ApplicationField.objects.get(id=field_id)
-
-        if not permission_manager.has_field_permission(
-            application_field,
-            create_permission_str(content_type.model_class(), "view")
-        ):
-            return HttpResponse("Permission denied", status=403)
-
-        toggle_field_visibility(request.user, content_type, field_id, view_type)
-    except (ValueError, ApplicationField.DoesNotExist) as e:
-        return HttpResponse(f"Invalid field: {e}", status=400)
-
-    return None
-
-
 def _render_data_view_body(
     request: HttpRequest,
     state: DataViewQueryState,
@@ -569,55 +443,3 @@ def data_view_action(request: HttpRequest, content_type_id: int, action: str) ->
     return definition.renderer_cls.handle_action(action, request, state)
     
 
-@router.register(
-    path="components/change_data_view_preference/<int:content_type_id>/",
-    name="components_change_data_view_preference",
-)
-def change_data_view_preference(request: HttpRequest, content_type_id: int) -> HttpResponse:
-    """Changes the datatable preference
-
-    Args:
-        request (HttpRequest): the request object
-        content_type_id (int): The content type ID
-
-    Returns:
-        HttpResponse: the rendered datatable with the different preferences
-    """
-    if request.method != "POST":
-        return HttpResponse("Method not allowed", status=405)
-    
-    # Get the content type, user, and list view preference
-    content_type = get_object_or_404(ContentType, id=content_type_id)
-    user = request.user
-    manager = PreferenceManager(user)
-    preference = manager.get_or_create_selected(UserListViewPreference, {
-        "content_type_id": content_type.id,
-    })
-    
-    if not manager.can_manage(preference):
-        return HttpResponse("Forbidden", status=403)
-    
-    operation = _get_preference_operation(request.POST)
-    match operation:
-        case "change_type":
-            error_response = _change_data_view_type(preference, request.POST)
-        case "split_view":
-            error_response = _change_split_view(preference, request.POST)
-        case "opt":
-            data_view_fields = get_data_view_fields(preference)
-            error_response = _change_data_view_options(
-                preference,
-                _get_accessible_application_fields(data_view_fields),
-                request.POST,
-            )
-        case "field":
-            error_response = _change_data_view_field_visibility(request, content_type, preference, request.POST)
-        case "default_filters":
-            error_response = _change_default_filters(preference, request.POST)
-        case _:
-            error_response = None
-
-    if error_response is not None:
-        return error_response
-    
-    return _render_display_options(request, content_type_id, preference)
